+++ capi/include/blazesym.h
@@ -991,6 +991,38 @@ typedef struct blaze_symbolize_src_elf {
 } blaze_symbolize_src_elf;
 
 /**
+ * Options influencing the address symbolization process.
+ */
+typedef struct blaze_symbolize_opts {
+  /**
+   * The size of this object's type.
+   *
+   * Make sure to initialize it to `sizeof(<type>)`. This member is used to
+   * ensure compatibility in the presence of member additions.
+   */
+  size_t type_size;
+  /**
+   * Whether or not to operate in permissive mode.
+   *
+   * In permissive mode batch address symbolization does not
+   * short-circuit on error but keeps going. Failure to symbolize an
+   * individual addresses will be communicated via the
+   * [`Unknown`][Symbolized::Unknown] variant of the corresponding
+   * [`Symbolized`] object.
+   *
+   * Note that this does not make symbolization infallible: failure
+   * of an operation affecting all addresses (such as opening the ELF
+   * file) will still be reported as regular error.
+   */
+  bool permissive;
+  /**
+   * Unused member available for future expansion. Must be initialized
+   * to zero.
+   */
+  uint8_t reserved[23];
+} blaze_symbolize_opts;
+
+/**
  * The parameters to load symbols and debug information from "raw" Gsym data.
  */
 typedef struct blaze_symbolize_src_gsym_data {
@@ -1447,6 +1479,31 @@ const struct blaze_syms *blaze_symbolize_elf_virt_offsets(blaze_symbolizer *symb
                                                           const struct blaze_symbolize_src_elf *src,
                                                           const uint64_t *virt_offsets,
                                                           size_t virt_offset_cnt);
+
+/**
+ * Symbolize virtual offsets in an ELF file, providing additional
+ * symbolization options.
+ *
+ * On success, the function returns a [`blaze_syms`] containing an
+ * array of `virt_offset_cnt` [`blaze_sym`] objects. The returned
+ * object should be released using [`blaze_syms_free`] once it is no
+ * longer needed.
+ *
+ * On error, the function returns `NULL` and sets the thread's last error to
+ * indicate the problem encountered. Use [`blaze_err_last`] to retrieve this
+ * error.
+ *
+ * # Safety
+ * - `symbolizer` needs to point to a valid [`blaze_symbolizer`] object
+ * - `src` needs to point to a valid [`blaze_symbolize_src_elf`] object
+ * - `virt_offsets` needs to point to an array of `virt_offset_cnt` addresses
+ * - `opts` needs to point to a valid `blaze_symbolize_opts` object
+ */
+const struct blaze_syms *blaze_symbolize_elf_virt_offsets_opts(blaze_symbolizer *symbolizer,
+                                                               const struct blaze_symbolize_src_elf *src,
+                                                               const uint64_t *virt_offsets,
+                                                               size_t virt_offset_cnt,
+                                                               const struct blaze_symbolize_opts *opts);
 
 /**
  * Symbolize file offsets in an ELF file.